Make code compile again
