+++ src/client/chat/systems.rs
@@ -43,7 +43,7 @@ pub fn send_messages_system(
 
         client.send_message(
             DefaultChannel::ReliableOrdered,
-            bincode::serialize(&NetworkingMessage::ChatMessageSend(message)).unwrap(),
+            bincode::serialize(&lib::NetworkingMessage::ChatMessageSend(message)).unwrap(),
         );
     }
 }
@@ -249,7 +249,7 @@ mod tests {
     use super::*;
     use bevy::ecs::event::Events;
     use chat_events::{ChatClearEvent, SingleChatSendEvent};
-    use rsmc::ChatMessage;
+    use crate::lib::ChatMessage;
 
     fn get_chat_messages(app: &mut App) -> Vec<String> {
         let mut messages = app
