+++ src/client/networking/mod.rs
@@ -4,7 +4,7 @@ use bevy_renet::{
     netcode::{ClientAuthentication, NetcodeClientPlugin, NetcodeClientTransport},
     RenetClientPlugin,
 };
-use rsmc::connection_config;
+use crate::lib::connection_config;
 
 use crate::prelude::*;
