+++ src/client/terrain/util/mod.rs
@@ -1,9 +1,5 @@
 pub mod blocks;
-pub mod buffer_serializer;
-pub mod chunk;
 pub mod mesher;
 
 pub use blocks::*;
-pub use buffer_serializer::*;
-pub use chunk::*;
 pub use mesher::*;
