+++ src/server/terrain/events.rs
@@ -3,5 +3,5 @@ use crate::prelude::*;
 #[derive(Event)]
 pub struct BlockUpdateEvent {
     pub position: Vec3,
-    pub block: BlockId,
+    pub block: lib::BlockId,
 }
