+++ src/server/terrain/util/blocks.rs
@@ -1,7 +1,7 @@
 use crate::prelude::*;
 
 pub struct Block {
-    pub id: BlockId,
+    pub id: lib::BlockId,
     pub is_solid: bool,
 }
 
@@ -15,18 +15,18 @@ macro_rules! add_block {
 }
 
 pub static BLOCKS: [Block; 14] = [
-    add_block!(BlockId::Air, false),
-    add_block!(BlockId::Grass, true),
-    add_block!(BlockId::Dirt, true),
-    add_block!(BlockId::Stone, true),
-    add_block!(BlockId::Bedrock, true),
-    add_block!(BlockId::RedSand, true),
-    add_block!(BlockId::BrownTerracotta, true),
-    add_block!(BlockId::CyanTerracotta, true),
-    add_block!(BlockId::GrayTerracotta, true),
-    add_block!(BlockId::LightGrayTerracotta, true),
-    add_block!(BlockId::OrangeTerracotta, true),
-    add_block!(BlockId::RedTerracotta, true),
-    add_block!(BlockId::Terracotta, true),
-    add_block!(BlockId::YellowTerracotta, true),
+    add_block!(lib::BlockId::Air, false),
+    add_block!(lib::BlockId::Grass, true),
+    add_block!(lib::BlockId::Dirt, true),
+    add_block!(lib::BlockId::Stone, true),
+    add_block!(lib::BlockId::Bedrock, true),
+    add_block!(lib::BlockId::RedSand, true),
+    add_block!(lib::BlockId::BrownTerracotta, true),
+    add_block!(lib::BlockId::CyanTerracotta, true),
+    add_block!(lib::BlockId::GrayTerracotta, true),
+    add_block!(lib::BlockId::LightGrayTerracotta, true),
+    add_block!(lib::BlockId::OrangeTerracotta, true),
+    add_block!(lib::BlockId::RedTerracotta, true),
+    add_block!(lib::BlockId::Terracotta, true),
+    add_block!(lib::BlockId::YellowTerracotta, true),
 ];
