+++ src/shared/buffer_serializer.rs
