+++ src/shared/prelude.rs
@@ -1,0 +1,1 @@
+use serde::{Deserialize, Serialize};
